refactor: update add schema modal to implement hooks pattern
